+++ frontend/ios/App/App/public/cordova_plugins.js
@@ -1,5 +1,5 @@
 
-  cordova.define('cordova/plugin_list', function(require, exports, module) {
+  window.cordova.define('cordova/plugin_list', function(require, exports, module) {
     module.exports = [
       
     ];